+++ src/data/codec/symbol/mod.rs
@@ -13,6 +13,7 @@ mod braille; // BrailleByte
 // mod ean;
 // mod pdf417;
 // mod qr;
+mod quadrant; // Quadrant4
 // mod upc;
 mod yijing; // YijingHexagram
 
@@ -21,11 +22,12 @@ crate::structural_mods! { // _mods
         pub use super::{
             // aztec::_all::*,
             braille::BrailleByte,
-            // code128::_all::*,
+            // code128::_all::G*
             // datamatrix::_all::*,
             // ean::_all::*,
             // pdf417::_all::*,
             // qr::_all::*,
+            quadrant::Quadrant4,
             // upc::_all::*,
             yijing::YijingHexagram,
         };